Remove HomeNewsPanelFooter
